Primera versión

+++ src/app/tasks/pages/edit-page/edit-page.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { TasksService } from '../../services/task.service';
 import { Task } from '../../interfaces/task.interface';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
+import { SuccessDialogComponent } from '../../components/dialogs/success-dialog/success-dialog.component';
+import { ConfirmDialogComponent } from '../../components/dialogs/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-edit-page',
@@ -12,13 +15,13 @@ export class EditPageComponent implements OnInit {
 
   availableIds: number[] = [];
 
-  constructor(private fb: FormBuilder, private taskService: TasksService) {}
+  constructor(private fb: FormBuilder, private taskService: TasksService, public dialog: MatDialog) {}
 
   public taskForm: FormGroup = this.fb.group({
     id: ['', Validators.required],
     estado: ['', Validators.required],
     personaAsignada: ['', [Validators.required, Validators.maxLength(20)]],
-    description: ['', [Validators.required, Validators.maxLength(50)]],
+    description: ['', [Validators.required, Validators.maxLength(100)]],
     dificultad: ['', Validators.required],
     categoria: ['', Validators.required]
   });
@@ -45,23 +48,32 @@ export class EditPageComponent implements OnInit {
       return;
     }
 
-    const task: Task = this.taskForm.value;
-    this.taskService.updateTask(task).subscribe({
-      next: response => {
-        console.log('Tarea actualizada:', response);
-        // Aquí puedes agregar lógica adicional, como redirigir al usuario o mostrar un mensaje de éxito
-      },
-      error: error => {
-        console.error('Error al actualizar la tarea:', error);
-        // Aquí puedes manejar el error, como mostrar un mensaje de error al usuario
-      },
-      complete: () => {
-        console.log('Operación completada');
-        // Aquí puedes agregar lógica adicional que se ejecutará cuando la operación se complete
-      }
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      data: { message: '¿Estás seguro de que deseas editar esta tarea?' }
     });
 
-    this.resetForm();
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        const task: Task = this.taskForm.value;
+        this.taskService.updateTask(task).subscribe({
+          next: response => {
+            console.log('Tarea actualizada:', response);
+            this.openSuccessDialog('Tarea editada correctamente con el ID: ' + response.id);
+          },
+          error: error => {
+            console.error('Error al actualizar la tarea:', error);
+            // Aquí puedes manejar el error, como mostrar un mensaje de error al usuario
+          },
+          complete: () => {
+            console.log('Operación completada');
+            // Aquí puedes agregar lógica adicional que se ejecutará cuando la operación se complete
+          }
+        });
+
+        this.resetForm();
+      }
+    });
   }
 
   onDelete(): void {
@@ -69,23 +81,32 @@ export class EditPageComponent implements OnInit {
       return;
     }
 
-    const id = this.taskForm.get('id')!.value;
-    this.taskService.deleteTask(id).subscribe({
-      next: () => {
-        console.log('Tarea eliminada');
-        // Aquí puedes agregar lógica adicional, como redirigir al usuario o mostrar un mensaje de éxito
-      },
-      error: error => {
-        console.error('Error al eliminar la tarea:', error);
-        // Aquí puedes manejar el error, como mostrar un mensaje de error al usuario
-      },
-      complete: () => {
-        console.log('Operación completada');
-        // Aquí puedes agregar lógica adicional que se ejecutará cuando la operación se complete
-      }
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      data: { message: '¿Estás seguro de que deseas eliminar esta tarea?' }
     });
 
-    this.resetForm();
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        const id = this.taskForm.get('id')!.value;
+        this.taskService.deleteTask(id).subscribe({
+          next: () => {
+            console.log('Tarea eliminada');
+            this.openSuccessDialog('Tarea eliminada correctamente.');
+          },
+          error: error => {
+            console.error('Error al eliminar la tarea:', error);
+            // Aquí puedes manejar el error, como mostrar un mensaje de error al usuario
+          },
+          complete: () => {
+            console.log('Operación completada');
+            // Aquí puedes agregar lógica adicional que se ejecutará cuando la operación se complete
+          }
+        });
+
+        this.resetForm();
+      }
+    });
   }
 
   private resetForm(): void {
@@ -104,6 +125,13 @@ export class EditPageComponent implements OnInit {
       control?.setErrors(null); // Eliminar errores de validación
       control?.markAsPristine();
       control?.markAsUntouched();
+    });
+  }
+
+  private openSuccessDialog(message: string): void {
+    this.dialog.open(SuccessDialogComponent, {
+      width: '250px',
+      data: { message }
     });
   }
 }
